requirements.txt
Flask
google-generativeai
requests
beautifulsoup4
gunicorn
Flask-Cors